# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Purchase Order System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/import:
    post:
      operationId: importCustomers
      tags: [customers]
      summary: Bulk import customers from a spreadsheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportCustomersBody"
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportCustomersResult"
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Customer deleted
  /customer-auth/me:
    get:
      operationId: getCustomerAuthMe
      tags: [customerAuth]
      summary: Get the currently logged-in customer
      responses:
        "200":
          description: Current customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSession"
        "401":
          description: Not logged in
  /customer-auth/login:
    post:
      operationId: customerLogin
      tags: [customerAuth]
      summary: Log in with customer number and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSession"
        "401":
          description: Invalid credentials or password not set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAuthError"
  /customer-auth/set-password:
    post:
      operationId: customerSetPassword
      tags: [customerAuth]
      summary: Set a password for a customer that has never logged in before
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSetPasswordBody"
      responses:
        "200":
          description: Password set and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSession"
        "400":
          description: Validation error, or email required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAuthError"
        "404":
          description: Customer number not found
  /customer-auth/forgot-password:
    post:
      operationId: customerForgotPassword
      tags: [customerAuth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent, or email needed before it can be sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerForgotPasswordResult"
        "404":
          description: Customer number not found
  /customer-auth/reset-password:
    post:
      operationId: customerResetPassword
      tags: [customerAuth]
      summary: Reset password using a reset token from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerResetPasswordBody"
      responses:
        "200":
          description: Password reset and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSession"
        "400":
          description: Invalid or expired reset token
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        customerNumber:
          type: integer
          nullable: true
        name:
          type: string
        billTo:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        primaryContact:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        shipTo:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateCustomerBody:
      type: object
      properties:
        customerNumber:
          type: integer
          nullable: true
        name:
          type: string
        billTo:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        primaryContact:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        shipTo:
          type: string
          nullable: true
      required: [name]
    UpdateCustomerBody:
      type: object
      properties:
        customerNumber:
          type: integer
          nullable: true
        name:
          type: string
        billTo:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        primaryContact:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        shipTo:
          type: string
          nullable: true
    ImportCustomersBody:
      type: object
      properties:
        customers:
          type: array
          items:
            type: object
            properties:
              customerNumber:
                type: integer
                nullable: true
              name:
                type: string
              billTo:
                type: string
                nullable: true
              email:
                type: string
                nullable: true
              primaryContact:
                type: string
                nullable: true
              phone:
                type: string
                nullable: true
              shipTo:
                type: string
                nullable: true
            required: [name]
      required: [customers]
    ImportCustomersResult:
      type: object
      properties:
        customersCreated:
          type: integer
      required: [customersCreated]
    CustomerSession:
      type: object
      properties:
        id:
          type: integer
        customerNumber:
          type: integer
          nullable: true
        name:
          type: string
        email:
          type: string
          nullable: true
      required: [id, name]
    CustomerAuthError:
      type: object
      properties:
        error:
          type: string
        needsSetup:
          type: boolean
        needsEmail:
          type: boolean
      required: [error]
    CustomerLoginBody:
      type: object
      properties:
        customerNumber:
          type: integer
        password:
          type: string
      required: [customerNumber, password]
    CustomerSetPasswordBody:
      type: object
      properties:
        customerNumber:
          type: integer
        password:
          type: string
        email:
          type: string
          nullable: true
      required: [customerNumber, password]
    CustomerForgotPasswordBody:
      type: object
      properties:
        customerNumber:
          type: integer
        email:
          type: string
          nullable: true
      required: [customerNumber]
    CustomerForgotPasswordResult:
      type: object
      properties:
        ok:
          type: boolean
        needsEmail:
          type: boolean
        message:
          type: string
      required: [ok]
    CustomerResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required: [token, password]
...